Improve backtrace output format.
