+++ src/elf_lib.c
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2015 Equinox Payments, LLC
+ *
+ * Permission to use, copy, modify, and distribute this software for any
+ * purpose with or without fee is hereby granted, provided that the above
+ * copyright notice and this permission notice appear in all copies.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ * ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ */
+
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
@@ -8,33 +24,35 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include "elf_lib.h"
 #include "util.h"
 #include "xmalloc.h"
 
-#define TARGET_ELF_CLASS	32
-
-#if TARGET_ELF_CLASS == 32
-typedef Elf32_Ehdr	Elf_Ehdr;
-typedef Elf32_Shdr	Elf_Shdr;
-typedef Elf32_Sym	Elf_Sym;
-#else
+#if TARGET_ELF_CLASS == 64
 typedef Elf64_Ehdr	Elf_Ehdr;
 typedef Elf64_Shdr	Elf_Shdr;
 typedef Elf64_Sym	Elf_Sym;
+#else
+typedef Elf32_Ehdr	Elf_Ehdr;
+typedef Elf32_Shdr	Elf_Shdr;
+typedef Elf32_Sym	Elf_Sym;
 #endif
 
 /*
  * An ELF file.
  */
 struct elf {
-	char		*path;	// Filesystem path to ELF file.
-	int		 fd;	// File descriptor of an open ELF file.
-	Elf_Ehdr	 hdr;	// Pointer to ELF header.
-	unsigned long	 shnum;	// Number of section headers.
+	char		*path;	/* Filesystem path to ELF file. */
+	int		 fd;	/* File descriptor of an open ELF file. */
+	Elf_Ehdr	 hdr;	/* Pointer to ELF header. */
+	unsigned long	 shnum;	/* Number of section headers. */
 };
 
+/*
+ * Read a block from ELF file, return true if successful.
+ */
 static bool
-read_struct(const struct elf *elf, off_t off, void *ptr, size_t size)
+read_block(const struct elf *elf, off_t off, void *ptr, size_t size)
 {
 	ssize_t	 n;
 
@@ -42,6 +60,9 @@ read_struct(const struct elf *elf, off_t off, void *ptr, size_t size)
 	return n == (ssize_t)size;
 }
 
+/*
+ * Read the ELF section header at the given index, return true if successful.
+ */
 static bool
 get_shdr_index(const struct elf *elf, Elf_Shdr *shp, unsigned long ndx)
 {
@@ -49,7 +70,7 @@ get_shdr_index(const struct elf *elf, Elf_Shdr *shp, unsigned long ndx)
 	Elf_Shdr	 sh;
 
 	off = elf->hdr.e_shoff + elf->hdr.e_shentsize * ndx;
-	if (!read_struct(elf, off, &sh, sizeof sh)) {
+	if (!read_block(elf, off, &sh, sizeof sh)) {
 		warningx("%s: failed to read section header %lu, at offset %lx", elf->path, ndx, (unsigned long)off);
 		return false;
 	}
@@ -57,6 +78,10 @@ get_shdr_index(const struct elf *elf, Elf_Shdr *shp, unsigned long ndx)
 	return true;
 }
 
+/*
+ * Read the first ELF section header which has the given type, return true
+ * if successful.
+ */
 static bool
 get_shdr_type(const struct elf *elf, Elf_Shdr *shp, unsigned long type)
 {
@@ -74,6 +99,9 @@ get_shdr_type(const struct elf *elf, Elf_Shdr *shp, unsigned long type)
 	return false;
 }
 
+/*
+ * Open an ELF file, return a handle if successful.
+ */
 struct elf *
 elf_open(const char *path)
 {
@@ -87,7 +115,7 @@ elf_open(const char *path)
 		goto error;
 	}
 	elf->path = xstrdup(path);
-	if (!read_struct(elf, 0, &elf->hdr, sizeof(elf->hdr))) {
+	if (!read_block(elf, 0, &elf->hdr, sizeof(elf->hdr))) {
 		warningx("%s: failed to read ELF header", path);
 		goto error;
 	}
@@ -99,7 +127,7 @@ elf_open(const char *path)
 
 	if (elf->hdr.e_shnum > 0) {
 		elf->shnum = elf->hdr.e_shnum;
-	} else if (!read_struct(elf, elf->hdr.e_shoff, &sh, sizeof sh)) {
+	} else if (!read_block(elf, elf->hdr.e_shoff, &sh, sizeof sh)) {
 		warningx("%s: failed to read first section header", path);
 		goto error;
 	}
@@ -114,6 +142,9 @@ error:
 	return NULL;
 }
 
+/*
+ * Close an ELF file.
+ */
 void
 elf_close(struct elf *elf)
 {
@@ -122,12 +153,19 @@ elf_close(struct elf *elf)
 	free(elf);
 }
 
+/*
+ * Return true if the ELF file is a shared object.
+ */
 bool
 elf_is_shared_object(const struct elf *elf)
 {
 	return elf->hdr.e_type == ET_DYN;
 }
 
+/*
+ * Search for a symbol which matches addr, put it's name into buf and offset
+ * from beginning into offp. Return true if successful.
+ */
 bool
 elf_resolve_sym(const struct elf *elf, unsigned long addr, char *buf, size_t bufsize, unsigned long *offp)
 {
@@ -153,7 +191,7 @@ nextsymtab:
 		for (off = symtabhdr.sh_offset;
 		    off < symtabhdr.sh_offset + symtabhdr.sh_size;
 		    off += symtabhdr.sh_entsize) {
-			if (!read_struct(elf, off, &sym, sizeof(sym))) {
+			if (!read_block(elf, off, &sym, sizeof(sym))) {
 				warningx("%s: failed to read symbol at offset %lu", elf->path, (unsigned long)off);
 				goto nextsymtab;
 			}
