+++ src/elf_lib.h
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2015 Equinox Payments, LLC
+ *
+ * Permission to use, copy, modify, and distribute this software for any
+ * purpose with or without fee is hereby granted, provided that the above
+ * copyright notice and this permission notice appear in all copies.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ * ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ */
+
+#ifndef ELF_LIB_H
+#define ELF_LIB_H
+
+#include <stdbool.h>
+
+struct elf;
+
+struct elf	*elf_open(const char *);
+void		 elf_close(struct elf *);
+bool		 elf_is_shared_object(const struct elf *);
+bool		 elf_resolve_sym(const struct elf *, unsigned long, char *, size_t, unsigned long *);
+
+#endif
