+++ src/unwind.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014 Equinox Payments, LLC
+ * Copyright (c) 2014, 2015 Equinox Payments, LLC
  *
  * Permission to use, copy, modify, and distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -657,6 +657,10 @@ guess_unwind(struct proc *p)
 	}
 }
 
+/*
+ * Use symbol tables in corresponding ELF files to resolve the names of the
+ * functions in the backtrace.
+ */
 static void
 resolve_func_names(struct proc *p)
 {
@@ -664,6 +668,7 @@ resolve_func_names(struct proc *p)
 	struct map	*map;
 	char		 name[128];
 	word_t		 addr;
+	unsigned long	 off;
 
 	TAILQ_FOREACH(frame, &p->backtrace, entry) {
 		*name = '\0';
@@ -676,8 +681,11 @@ resolve_func_names(struct proc *p)
 				frame->map = map;
 				if (elf_is_shared_object(map->elf))
 					addr -= map->start;
-				if (elf_resolve_sym(map->elf, addr, name, sizeof name, &frame->func.off))
+				frame->func.addr = addr;
+				if (elf_resolve_sym(map->elf, addr, name, sizeof name, &off)) {
 					frame->func.name = xstrdup(name);
+					frame->func.off = (word_t)off;
+				}
 				break;
 			}
 		}
